% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_add.zeros.commonName.R
\name{add.zeros.commonName}
\alias{add.zeros.commonName}
\title{prepare data with common names}
\usage{
add.zeros.commonName(pc2)
}
\arguments{
\item{df}{A data frame object}
}
\value{
A df
}
\description{
Creates a dataframe
}
\examples{
add.zeros.commonName(df)

}
