% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.zeros.count.r
\name{add.zeros.count}
\alias{add.zeros.count}
\title{Adds zeros to dataframe for statistical analysis}
\usage{
add.zeros.count(pc)
}
\arguments{
\item{df}{A data frame object}
}
\value{
A data frame
}
\description{
Takes bird data that has been prepared by bird_prepare function and turned into the same information but with zeros where no detections were found.
}
\examples{
df2 = add.zeros.count(df)

}
