% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.zeros.noCount.r
\name{add.zeros.noCount}
\alias{add.zeros.noCount}
\title{Adds zeros to dataframe for statistical analysis}
\usage{
add.zeros.noCount(pc)
}
\arguments{
\item{df}{A data frame object}
}
\value{
A data frame
}
\description{
Takes bird data that has been prepared by bird_prepare function and turned into the same information but with zeros where no detections were found.
}
\examples{
df2 = bird_species.common(pc)

}
