% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_remove.first.r, R/bird_remove_first.r
\name{all_remove.first}
\alias{all_remove.first}
\title{Removes first row of df if it is missing.}
\usage{
all_remove.first(df)

all_remove.first(df)
}
\arguments{
\item{df}{A dataframe object}
}
\value{
A data frame without empty first row

A data frame without a blank first row.
}
\description{
CADC will add an empty row when downloading data.

A reoccuring problem when importing data frames into rstudio is a blank first row that is added to the data frame. This function looks at the firsts row to see if the first cell has an empty value. If it does then it will removes the row and return the dataset.
}
\examples{
data = all_remove.first(data)

data = remove_first(data)

}
