% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_focal.Abundance.Facet.R
\name{bird_focal.Abundance.Facet}
\alias{bird_focal.Abundance.Facet}
\title{Graph for species abundance by focal group}
\usage{
bird_focal.Abundance.Facet(df, relative = T, commonName = TRUE,
  distance, transect = c(levels(as.factor(df$Transect))),
  surveyyear = c(levels(as.factor(df$YEAR))),
  choose_focal_group = c("Grassland", "Oak.Woodland", "Riparian",
  "Mountian.Meadow"), removeSpecies = c())
}
\arguments{
\item{df}{A data frame object}

\item{relative}{boolian: do you want to account for effort}

\item{commonName}{if you want common names}

\item{distance}{pick a distance to select by}

\item{transect}{the transect}

\item{surveyyear}{the year}

\item{choose_focal_group}{focal groups}

\item{removeSpecies}{species you want removed.}
}
\value{
A graph
}
\description{
Creates a graph of species abundance by focal group
}
\examples{
bird_focal.Abundance.Facet(df)

}
