% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_focal.graph_Abundance.R
\name{bird_focal.graph_Abundance}
\alias{bird_focal.graph_Abundance}
\title{Graph for species abundance by focal group}
\usage{
bird_focal.graph_Abundance(df, relative, commonName = FALSE,
  distance = 300, transect = c(levels(as.factor(df$Transect))),
  surveyyear = c(levels(as.factor(df$YEAR))),
  choose_focal_group = c("Grassland", "Oak.Woodland", "Riparian",
  "Mountian.Meadow"))
}
\arguments{
\item{df}{A data frame object}

\item{relative}{boolian: do you want to account for effort}

\item{distance}{pick a distance to select by}

\item{transect}{the transect}

\item{surveyyear}{the year}

\item{choose_focal_group}{focal groups}

\item{common}{if you want common names}
}
\value{
A graph
}
\description{
Creates a graph of species abundance by focal group
}
\examples{
bird_focal.graph_Abundance(df)

}
