% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_focal.groups.r
\name{bird_focal.groups}
\alias{bird_focal.groups}
\title{Create a Focal Species Abundance Graph. Bar for each: Grassland, OakWoodland, and Riparian Spp}
\usage{
bird_focal.groups(
  df2,
  distance_focal,
  transect = c(levels(as.factor(df$Transect))),
  surveyyear = c(levels(as.factor(df$YEAR)))
)
}
\arguments{
\item{df2}{A data frame object. Only will take bird_prepare(). See those functions for additional previous steps.}

\item{distance_focal}{Filters the data by selecting only detections less than this value.}

\item{transect}{A ranch code or a list of ranch codes ie."TOKA".}

\item{surveyyear}{A year or multiple years. ie. c(2016,2018)}
}
\value{
A graph that summerises focal species abundance.
}
\description{
Takes bird data that has been prepared by bird_prepare function, to create a focal species abundance per year graph.
}
\examples{
bird_focal.group(df2, transect="TOKA", surveyyear= 2016, distance_focal=300)

}
