% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_species.pointYear.r
\name{bird_species.pointYear}
\alias{bird_species.pointYear}
\title{Creates Species dataframe summarized by point and year.}
\usage{
bird_species.pointYear(
  df,
  distance = 300,
  transect = c(levels(as.factor(df$Transect))),
  surveyyear = c(levels(as.factor(df$YEAR)))
)
}
\arguments{
\item{df}{A data frame object}
}
\value{
A data frame
}
\description{
Takes bird data that has been prepared by bird_prepare function and creates a dataframe summerising all observations of species by year.
}
\examples{
data = bird_species.pointYear(df)

}
