% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_update.species.names.r
\name{bird_updateSpeciesNames}
\alias{bird_updateSpeciesNames}
\title{Update species names.}
\usage{
bird_updateSpeciesNames(df)
}
\arguments{
\item{df}{A data frame object}
}
\value{
A data frame that is subsetted based on detection
}
\description{
changes the species list names to change WESJ to CASJ.
}
\examples{
data = bird_updateSpeciesNames(df)

}
