% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_numberOfPoint.R
\name{numberOfPoints}
\alias{numberOfPoints}
\title{Prepared data for by getting number of points}
\usage{
numberOfPoints(df, common)
}
\arguments{
\item{df}{is a dataframe. Only works with newpc2 created by add.zeros() and it's previous steps.}

\item{common}{This defines if you want to add in common names of birds to the dataframe.}
}
\value{
Data frame with effort per ranch per year
}
\description{
This gets the amount of effort per ranch per year
}
\examples{
numberOfPoints(df, common=T)

}
