% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareBirdRichnessData.R
\name{prepareBirdRichnessData}
\alias{prepareBirdRichnessData}
\title{Prepared data for visualization}
\usage{
prepareBirdRichnessData(df, distance,
  transect = c(levels(as.factor(df$Transect))),
  surveyyear = c(levels(as.factor(df$YEAR))))
}
\arguments{
\item{df}{is a dataframe. Only works with newpc2 created by add.zeros() and it's previous steps.}

\item{distance}{is a dataframe. Only works with newpc2 created by add.zeros() and it's previous steps.}

\item{transect}{is a dataframe. Only works with newpc2 created by add.zeros() and it's previous steps.}

\item{surveyyear}{is a dataframe. Only works with newpc2 created by add.zeros() and it's previous steps.}
}
\value{
Data frame with the is the wide version of newpc2 with added column of Richness
}
\description{
This function turns the dataframe that prepares the data for analysis and graphs
}
\examples{
prepareBirdRichnessData(df)

}
