% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bird_prepareData_R_Abundance.R
\name{prepareData_R_Abundance}
\alias{prepareData_R_Abundance}
\title{Summarize data for analysis}
\usage{
prepareData_R_Abundance(df, NumberOfPoints = NumberOfPoints, common,
  distance)
}
\arguments{
\item{df}{A data frame object}

\item{NumberOfPoints}{A data frame object}

\item{common}{if you want common names}

\item{distance}{pick a distance to select by}
}
\value{
A data frame with visits
}
\description{
Takes data and combines with numberOfPoints to get the right effort.
}
\examples{
data = prepareData_R_Abundance(df)

}
